requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
discord.py==2.2.2
dropbox==11.36.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
ply==3.11
redis==4.5.4
requests==2.28.2
six==1.16.0
spotipy==2.23.0
stone==3.3.1
urllib3==1.26.15
yarl==1.9.1
